Move require tests from interpreter to kernel extn
